feat: made packageBanner a named export
