+++ .eslintignore
@@ -1,1 +1,1 @@
-commonjs/index.js
+commonjs/*.js
